+++ src/Jobs/MoveDirectoryStructure.php
@@ -25,7 +25,7 @@ class MoveDirectoryStructure implements ShouldQueue
 
     public function __construct(protected int $directoryId, protected int $newParentId, protected int $userId)
     {
-        $this->queue = 'dam-bulk';
+        $this->queue = config('dam.queues.bulk');
     }
 
     public function handle(): void